refactor: move surface_geom.rs to src/surface/geom.rs
