+++ src/surface.rs
@@ -1,6 +1,9 @@
 //! Shared terminal-surface identity: the `OzmaTerminal` marker and the
 //! render-bundle injection observer, which fire for every surface — tmux
-//! panes and the Default-mode shell alike.
+//! panes and the Default-mode shell alike. Surface geometry helpers live in
+//! `geom`.
+
+pub(crate) mod geom;
 
 use bevy::prelude::*;
 use ozma_tty_renderer::material::TerminalUiMaterial;
